backend/requirements.txt
kafka-python
pandas
requests
pyspark